pyproject.toml
[project]
name = "match"
version = "0.1"
requires-python = ">= 3.13"
dependencies = [
    "alembic==1.16.5",
    "black==24.10",
    "fastapi[standard]==0.115.6",
    "isort==5.13.2",
    "mypy==1.14.1",
    "pydantic==2.10.5",
    "pytest==8.3.4",
    "python-dotenv==1.0.1",
    "sentry-sdk[fastapi]==2.22.0",
    "sqlalchemy==2.0.43",
]

[tool.pytest.ini_options]
testpaths = ["match/tests"]


[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
exclude = [
    "tests"
]

[tool.isort]
profile = "black"
# isort doesn't support 3.13
py_version = 312
line_length = 100
